move function to get nutrition data to separate module
